minor fixes

+++ modules/task_3/ershov_a_deicstra_algorithm/deicstra_mpi.cpp
@@ -26,8 +26,8 @@ std::vector<int> getDeicstra(const std::vector<std::vector<int>>& graf,
   std::vector<bool> visitedTops(count);
   std::vector<int> dist(count, 10000);
   dist[top] = 0;
-  size_t min_dist = 0;
-  size_t min_vertex = top;
+  int min_dist = 0;
+  int min_vertex = top;
 
   while (min_dist < 10000) {
     size_t i = min_vertex;
